requirements.txt
aws-cdk-lib>=2.30.0
constructs>=10.0.0
PyYAML>=5.3.1
flake8>=4.0.1
